Dynamic Overview Table and testing pic link
There are some components that are still hard coded, but over all most
of the mechanics at least are tested at this point.
The pop-up control will need to be seeded better and the current data is
just a test set.
Tinkered with how the color scheme and episode prevalence looks. There
will need to be a more detailed episode check to correctly identify
based upon, probably name. Unsure as data has not been truly finalized.

Lastly was a test to see if the img link works without local hosting.