Voice actor fixes and overview updates
Fixes made to data files regarding voice actors. Applies to character
file and voice actor file.

Over view updated so that rows are limited to one line of text.
On mouse over creates a "highlighted" row column. But the colors are
really bad and need to be changed.
Disabled character picture for now.