Created a location datafile
Sorry, had to rush this one to prepare for lab today.
SLocation.js is of format [name, [episodenames ] ]

Currently two support functions.
One will grab all the locations that an episode has in it.
getLocationsInEpisode( takes name of episode)

The other will give a true/false on whether or not a location is inside
of an episode.
isLocationInEpisode(loc, nameOfEpisode)

loc can actually be a number or a string.
if a number then it will check based off of that location index.

episode must be the name of the episode.